# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Hush API
  version: 1.0.0
  description: >
    Contract for Hush Apple clients, Agent backend, and Unified Inbox.
    Quest steps are canonical in the fixed content library; the backend primarily returns quest_id.
servers:
  - url: https://api.example.com
    description: Demo
  - url: http://localhost:3000
    description: Local
...
paths:
...
  /v1/rest/evaluate:
    post:
      tags: [Rest]
      operationId: evaluateRest
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/ClientVersion"
        - $ref: "#/components/parameters/ContractVersion"
        - $ref: "#/components/parameters/DemoToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/usage-summary.schema.json"
      responses:
        "200":
          description: Rest suggestion
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            X-Contract-Version:
              $ref: "#/components/headers/ContractVersion"
            X-Hush-Data-Origin:
              $ref: "#/components/headers/DataOrigin"
          content:
            application/json:
              schema:
                $ref: "./schemas/rest-suggestion.schema.json"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /v1/rest/check-in:
    post:
      tags: [Rest]
      operationId: reflectFatigue
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/ClientVersion"
        - $ref: "#/components/parameters/ContractVersion"
        - $ref: "#/components/parameters/DemoToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/fatigue-check-in.schema.json"
      responses:
        "200":
          description: Fatigue reflection; at most one follow-up
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            X-Contract-Version:
              $ref: "#/components/headers/ContractVersion"
            X-Hush-Data-Origin:
              $ref: "#/components/headers/DataOrigin"
          content:
            application/json:
              schema:
                $ref: "./schemas/fatigue-reflection.schema.json"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /v1/rest/recommend:
    post:
      tags: [Rest]
      operationId: recommendQuest
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/ClientVersion"
        - $ref: "#/components/parameters/ContractVersion"
        - $ref: "#/components/parameters/DemoToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/rest-recommendation.schema.json#/$defs/RestRecommendationRequest"
      responses:
        "200":
          description: Quest ID selected from fixed content
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            X-Contract-Version:
              $ref: "#/components/headers/ContractVersion"
            X-Hush-Data-Origin:
              $ref: "#/components/headers/DataOrigin"
          content:
            application/json:
              schema:
                $ref: "./schemas/rest-recommendation.schema.json#/$defs/RestQuestRecommendation"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /v1/rest/feedback:
    post:
      tags: [Rest]
      operationId: recordRestFeedback
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/ClientVersion"
        - $ref: "#/components/parameters/ContractVersion"
        - $ref: "#/components/parameters/DemoToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/rest-feedback.schema.json"
      responses:
        "202":
          description: Accepted
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            X-Contract-Version:
              $ref: "#/components/headers/ContractVersion"
            X-Hush-Data-Origin:
              $ref: "#/components/headers/DataOrigin"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /v1/handoff/start:
    post:
      tags: [Handoff]
      operationId: startHandoff
      parameters:
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/ClientVersion"
        - $ref: "#/components/parameters/ContractVersion"
        - $ref: "#/components/parameters/DemoToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/handoff-start-request.schema.json"
      responses:
        "202":
          description: Handoff job created or existing idempotent job returned
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            X-Contract-Version:
              $ref: "#/components/headers/ContractVersion"
            X-Hush-Data-Origin:
              $ref: "#/components/headers/DataOrigin"
          content:
            application/json:
              schema:
                $ref: "./schemas/handoff-job.schema.json#/$defs/HandoffJob"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /v1/handoff/{jobId}:
    get:
      tags: [Handoff]
      operationId: getHandoffJob
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/ClientVersion"
        - $ref: "#/components/parameters/ContractVersion"
        - $ref: "#/components/parameters/DemoToken"
      responses:
        "200":
          description: Current state
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            X-Contract-Version:
              $ref: "#/components/headers/ContractVersion"
            X-Hush-Data-Origin:
              $ref: "#/components/headers/DataOrigin"
          content:
            application/json:
              schema:
                $ref: "./schemas/handoff-job.schema.json#/$defs/HandoffJobState"
        "404":
          $ref: "#/components/responses/Error"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /v1/handoff/{jobId}/cancel:
    post:
      tags: [Handoff]
      operationId: cancelHandoff
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/RequestId"
        - $ref: "#/components/parameters/ClientVersion"
        - $ref: "#/components/parameters/ContractVersion"
        - $ref: "#/components/parameters/DemoToken"
      responses:
        "202":
          description: Cancellation requested
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestId"
            X-Contract-Version:
              $ref: "#/components/headers/ContractVersion"
            X-Hush-Data-Origin:
              $ref: "#/components/headers/DataOrigin"
        "404":
          $ref: "#/components/responses/Error"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    AppBearer:
      type: http
      scheme: bearer
      description: Hush App session token. Cannot ingest connector events.
    ConnectorBearer:
      type: http
      scheme: bearer
      description: Connector Host token. Cannot query items or send drafts.
    DemoTokenAuth:
      type: apiKey
      in: header
      name: X-Hush-Demo-Token
      description: Available only when HUSH_DEMO_MODE=true.
  parameters:
    RequestId:
      name: X-Request-ID
      in: header
      required: true
      description: Must equal request_id in JSON request bodies.
      schema:
        type: string
        minLength: 1
    ClientVersion:
      name: X-Client-Version
      in: header
      required: true
      schema:
        type: string
        minLength: 1
    ContractVersion:
      name: X-Contract-Version
      in: header
      required: true
      schema:
        type: string
        const: "1.0"
    DemoToken:
      name: X-Hush-Demo-Token
      in: header
      required: false
      description: Enables Sample Mode only when the server also has HUSH_DEMO_MODE=true.
      schema:
        type: string
        minLength: 8
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^(?!\s)(?!.*\s$)[^\x00-\x1F\x7F-\x9F]+$'
  headers:
    RequestId:
      schema:
        type: string
    ContractVersion:
      schema:
        type: string
        const: "1.0"
    DataOrigin:
      schema:
        type: string
        enum: [real, mock, cached]
  responses:
    Error:
      description: Error
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestId"
        X-Contract-Version:
          $ref: "#/components/headers/ContractVersion"
        X-Hush-Data-Origin:
          $ref: "#/components/headers/DataOrigin"
      content:
        application/json:
          schema:
            $ref: "./schemas/error-response.schema.json"